nswiri_flask_api/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.3.3
